code/requirements.txt
akismet==1.3
Flask==3.0.2
Flask_Cors==4.0.0
notion_client==2.2.1
python-dotenv==1.0.1
